update code to use renamed directories / scripts / modules
