+++ ros_buildfarm/templates/doc/doc_job.xml.em
@@ -47,7 +47,7 @@ but disabled since the package is blacklisted (or not whitelisted) in the config
 @(SNIPPET(
     'scm',
     repo_spec=doc_repo_spec,
-    path='catkin_workspace/src/%s' % doc_repo_spec.name,
+    path='ws/src/%s' % doc_repo_spec.name,
     git_ssh_credential_id=git_ssh_credential_id,
 ))@
   <scmCheckoutRetryCount>2</scmCheckoutRetryCount>
@@ -186,7 +186,7 @@ else:
         ' -v $WORKSPACE/rosdoc_lite:/tmp/rosdoc_lite:ro' +
         ' -v $WORKSPACE/catkin-sphinx:/tmp/catkin-sphinx:ro' +
         ' -v $WORKSPACE/rosdoc_index:/tmp/rosdoc_index:ro' +
-        ' -v $WORKSPACE/catkin_workspace:/tmp/catkin_workspace' +
+        ' -v $WORKSPACE/ws:/tmp/ws' +
         ' -v $WORKSPACE/generated_documentation:/tmp/generated_documentation' +
         ' -v $WORKSPACE/docker_doc:/tmp/docker_doc' +
         hgcache_mount_arg +
@@ -222,7 +222,7 @@ else:
         ' -v $WORKSPACE/rosdoc_lite:/tmp/rosdoc_lite:ro' +
         ' -v $WORKSPACE/catkin-sphinx:/tmp/catkin-sphinx:ro' +
         ' -v $WORKSPACE/rosdoc_index:/tmp/rosdoc_index:ro' +
-        ' -v $WORKSPACE/catkin_workspace:/tmp/catkin_workspace' +
+        ' -v $WORKSPACE/ws:/tmp/ws' +
         ' -v $WORKSPACE/generated_documentation:/tmp/generated_documentation' +
         ' doc.%s_%s' % (rosdistro_name, doc_repo_spec.name.lower()),
         'echo "# END SECTION"',
@@ -233,9 +233,9 @@ else:
     script='\n'.join([
         'if [ "$skip_cleanup" = "false" ]; then',
         'echo "# BEGIN SECTION: Clean up to save disk space on agents"',
-        'rm -fr catkin_workspace/build_isolated',
-        'rm -fr catkin_workspace/devel_isolated',
-        'rm -fr catkin_workspace/install_isolated',
+        'rm -fr ws/build_isolated',
+        'rm -fr ws/devel_isolated',
+        'rm -fr ws/install_isolated',
         'rm -fr catkin-sphinx',
         'rm -fr rosdoc_index',
         'rm -fr rosdoc_lite',
