+++ ros_buildfarm/templates/doc/doc_task.Dockerfile.em
@@ -82,7 +82,7 @@ cmd = 'PYTHONPATH=/tmp/ros_buildfarm:$PYTHONPATH python3 -u' + \
     ' --rosdistro-name ' + rosdistro_name + \
     ' --os-code-name ' + os_code_name + \
     ' --arch ' + arch + \
-    ' --workspace-root /tmp/catkin_workspace' + \
+    ' --workspace-root /tmp/ws' + \
     ' --rosdoc-lite-dir /tmp/rosdoc_lite' + \
     ' --catkin-sphinx-dir /tmp/catkin-sphinx' + \
     ' --rosdoc-index /tmp/rosdoc_index' + \
