+++ ros_buildfarm/templates/prerelease/prerelease_clone_underlay_script.sh.em
@@ -11,6 +11,6 @@ fi
 
 @(TEMPLATE(
     'devel/devel_script_clone.sh.em',
-    workspace_path='catkin_workspace',
+    workspace_path='ws',
     scms=scms,
 ))@
