+++ scripts/devel/build_and_test.py
@@ -19,15 +19,15 @@ import os
 import sys
 
 from ros_buildfarm.catkin_workspace import call_catkin_make_isolated
-from ros_buildfarm.catkin_workspace import clean_workspace
-from ros_buildfarm.catkin_workspace import ensure_workspace_exists
 from ros_buildfarm.common import Scope
+from ros_buildfarm.workspace import clean_workspace
+from ros_buildfarm.workspace import ensure_workspace_exists
 
 
 def main(argv=sys.argv[1:]):
     parser = argparse.ArgumentParser(
-        description="Invoke 'catkin_make_isolated' on a workspace while "
-                    "enabling and running the tests")
+        description='Invoke the build tool on a workspace while enabling and '
+                    'running the tests')
     parser.add_argument(
         '--rosdistro-name',
         required=True,
